Added a test of inactive characters fading out
Check the debug scene to see my test implementation of fading out characters who aren't speaking, either by graying them out or by making them partially transparent.

My way of doing this was to pick one neutral sprite for each character, then define an image in Ren'Py that applies an effect to it to make it grayscale or transparent. Then for each line the characters spoke, I had them temporarily shift to their expression. If I want them to stay in a specific expression when not speaking, I'll have to define an image of that expression that's gray too.